Moved Command Action to ColorDetector.ts to perserve original file structure
